requirements.txt
Flask
pyBarcode
Pillow
wtforms
flask_wtf
